Add files via upload
This program reads a weighted undirected graph from stdin and writes the structure of the MST to stdout. The input will consist of a set of alphanumeric (string) vertex identifiers on one line, separated by spaces. The vertices will be followed by a list of edges, one edge per line, specified by two vertex identifiers and a numerical weight.